test(deep_causality_ast): Improved test organization.
