+++ deep_causality_ast/tests/const_tree/clone_tests.rs
@@ -1,0 +1,14 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
+ */
+
+use deep_causality_ast::ConstTree;
+
+#[test]
+fn test_clone_is_cheap() {
+    let original = ConstTree::new(10);
+    let cloned = original.clone();
+    assert!(original.ptr_eq(&cloned));
+    assert_eq!(*original.value(), *cloned.value());
+}
