+++ deep_causality_ast/tests/const_tree/default_tests.rs
@@ -1,0 +1,13 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
+ */
+
+use deep_causality_ast::ConstTree;
+
+#[test]
+fn test_default() {
+    let tree: ConstTree<i32> = ConstTree::default();
+    assert_eq!(*tree.value(), 0);
+    assert!(tree.is_leaf());
+}
