+++ deep_causality_ast/tests/const_tree/mod.rs
@@ -3,5 +3,15 @@
  * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 
-#[cfg(test)]
-mod const_tree_tests;
+mod accessors_tests;
+mod api_tests;
+mod clone_tests;
+mod debug_tests;
+mod default_tests;
+mod display_tests;
+mod from_tests;
+mod into_iter_tests;
+mod iter_tests;
+mod map_tests;
+mod mod_tests;
+mod partial_eq_tests;
