+++ deep_causality_tensor/src/types/causal_tensor/from/mod.rs
@@ -8,15 +8,14 @@ use crate::CausalTensor;
 impl<T: Clone> From<T> for CausalTensor<T> {
     /// Creates a scalar tensor (0-dimensional) from a single value.
     fn from(item: T) -> Self {
-        CausalTensor::new(vec![item], vec![]).expect("Failed to create scalar CausalTensor from T")
+        CausalTensor::from_vec_and_shape_unchecked(vec![item], &[])
     }
 }
 
 impl<'a, T: Clone> From<&'a T> for CausalTensor<T> {
     /// Creates a scalar tensor (0-dimensional) from a reference to a single value.
     fn from(item: &'a T) -> Self {
-        CausalTensor::new(vec![item.clone()], vec![])
-            .expect("Failed to create scalar CausalTensor from &T")
+        CausalTensor::from_vec_and_shape_unchecked(vec![item.clone()], &[])
     }
 }
